enh: Separate Python modules with functions for batch job submission [Applications]
