+++ Applications/lib/python/mirtk/atlas/spatiotemporal.py
@@ -23,12 +23,15 @@ import math
 import csv
 import json
 import re
-import errno
 import shutil
 import subprocess
 import time
-import mirtk
 from datetime import datetime
+
+import mirtk
+from mirtk.utils import makedirs
+from mirtk.batch.slurm  import submit as sbatch, wait as swait
+from mirtk.batch.condor import submit as cbatch, wait as cwait
 
 
 ##############################################################################
@@ -1477,418 +1480,6 @@ class SpatioTemporalAtlas(object):
 
 
 ##############################################################################
-# Batch execution using SLURM
-
-# ----------------------------------------------------------------------------
-def sbatch(name, command=None, args=[], opts={}, script=None, tasks=0, deps=[], logdir=None, log=None, queue='long', threads=8, memory=8 * 1024, workdir=None, verbose=1):
-    if threads <= 0:
-        raise ValueError("Must specify number of threads when executing as SLURM batch job!")
-    if script:
-        if command:
-            raise ValueError("Keyword arguments 'command' and 'script' are mutually exclusive")
-        shexec = "#!/bin/bash\nexec {0} <(cat <<END_OF_SCRIPT\n".format(sys.executable)
-        shexec += script.format(**opts)
-        shexec += "\nEND_OF_SCRIPT)"
-        if tasks > 0:
-            shexec += " $SLURM_ARRAY_TASK_ID"
-        shexec += "\n"
-        script = shexec
-    else:
-        script = "#!/bin/bash\n"
-        script += "\"{0}\" -c 'import sys; import socket;".format(sys.executable)
-        script += " sys.stdout.write(\"Host: \" + socket.gethostname() + \"\\n\\n\");"
-        script += " sys.path.insert(0, \"{0}\");".format(os.path.dirname(os.path.dirname(mirtk.__file__)))
-        script += " import mirtk; mirtk.check_call([\"{0}\"] + sys.argv[1:])".format(command if command else name)
-        script += "'"
-        for arg in args:
-            arg = str(arg)
-            if ' ' in arg:
-                arg = '"' + arg + '"'
-            script += ' ' + arg
-        for opt in opts:
-            arg = opts[opt]
-            if opt[0] != '-':
-                opt = '-' + opt
-            script += ' ' + opt
-            if arg is not None:
-                if isinstance(arg, (list, tuple)):
-                    arg = ' '.join([str(x) for x in arg])
-                else:
-                    arg = str(arg)
-                script += ' ' + arg
-        script += " -threads {0}".format(threads)
-        script += "\n"
-    argv = [
-        'sbatch',
-        '-J', name,
-        '-n', '1',
-        '-c', str(threads),
-        '-p', queue,
-        '--mem={0}M'.format(memory)
-    ]
-    if tasks > 0:
-        argv.append('--array=0-{}'.format(tasks - 1))
-    if logdir or log:
-        if not logdir:
-            logdir = os.path.dirname(log)
-        elif not log:
-            log = os.path.join(logdir, name)
-            if tasks > 0:
-                log += "_%A.%a.log"
-            log += "_%j.log"
-        if not os.path.exists(logdir):
-            os.makedirs(logdir)
-        argv.extend(['-o', log, '-e', log])
-    if deps:
-        if isinstance(deps, int):
-            deps = [deps]
-        deps = [str(dep) for dep in deps if dep > 0]
-        if deps:
-            argv.append('--dependency=afterok:' + ',afterok:'.join(deps))
-    if workdir:
-        argv.append('--workdir=' + os.path.abspath(workdir))
-    proc = subprocess.Popen(argv, stdout=subprocess.PIPE, stderr=subprocess.PIPE, stdin=subprocess.PIPE)
-    (out, err) = proc.communicate(input=script.encode('utf-8'))
-    if proc.returncode != 0:
-        raise Exception(err)
-    match = re.match('Submitted batch job ([0-9]+)', out)
-    if not match:
-        raise Exception("Failed to determine job ID from sbatch output:\n" + out)
-    jobid = int(match.group(1))
-    if verbose > 0:
-        if tasks > 0:
-            print("  Submitted batch {} (JobId={}, Tasks={})".format(name, jobid, tasks))
-        else:
-            print("  Submitted job {} (JobId={})".format(name, jobid))
-    return jobid
-
-
-# ----------------------------------------------------------------------------
-def swait(jobs, max_time=0, interval=60, verbose=0):
-    """Wait for SLURM jobs to finish."""
-    if not isinstance(jobs, (list, tuple)):
-        jobs = [jobs]
-    jobs = [job for job in jobs if job > 0]
-    num_wait = len(jobs)
-    if num_wait == 0:
-        return True
-    num_fail = 0
-    total_time = 0
-    re_state = re.compile("JobState=([A-Z]+)")
-    jobs = [str(job) for job in jobs]
-    iterations = 0
-    while num_wait > 0 and (max_time <= 0 or total_time < max_time):
-        time.sleep(interval)
-        total_time += interval
-        done = []
-        fail = []
-        batch = []
-        table = subprocess.check_output(["sacct", "--brief", "--parsable", "--jobs={}".format(','.join(jobs))])
-        for line in table.splitlines():
-            cols = line.split("|")
-            if cols[0] in jobs:
-                if cols[1] == "COMPLETED":
-                    done.append(cols[0])
-                elif cols[1] not in ("PENDING", "SUSPENDED", "RUNNING"):
-                    fail.append(cols[0])
-            elif cols[0].endswith(".batch"):
-                batch.append(cols[0][0:-6])
-        num_jobs = len(jobs)
-        num_done = 0
-        num_fail = 0
-        for job in batch:
-            try:
-                info = subprocess.check_output(["scontrol", "show", "job", job])
-                for line in info.splitlines():
-                    match = re_state.search(line)
-                    if match:
-                        num_jobs += 1
-                        if match.group(1) == "COMPLETED":
-                            num_done += 1
-                        elif match.group(1) not in ("PENDING", "SUSPENDED", "RUNNING"):
-                            num_fail += 1
-                num_jobs -= 1
-                try:
-                    done.remove(job)
-                except ValueError:
-                    pass
-                try:
-                    fail.remove(job)
-                except ValueError:
-                    pass
-            except subprocess.CalledProcessError:
-                pass  # scontrol forgets about no longer queued/running jobs
-        num_done += len(done)
-        num_fail += len(fail)
-        num_wait = num_jobs - num_done - num_fail
-        if verbose > 0 and (num_wait <= 0 or iterations % verbose == 0):
-            sys.stdout.write("{:%Y-%b-%d %H:%M:%S}".format(datetime.now()))
-            sys.stdout.write(" WAIT {} job(s) running/suspended/pending".format(num_wait))
-            if num_fail > 0:
-                sys.stdout.write(", {} failed".format(num_fail))
-            sys.stdout.write("\n")
-            sys.stdout.flush()
-        ++iterations
-    if num_wait > 0 and max_time > 0 and total_time >= max_time:
-        raise Exception("Exceeded maximum time waiting for jobs to complete!")
-    if total_time > 0:
-        time.sleep(10)  # wait a bit for files to be available from all NFS clients
-    return num_fail == 0
-
-
-##############################################################################
-# Batch execution using HTCondor
-
-# ----------------------------------------------------------------------------
-def cbatch(name, command=None, args=[], opts={}, script=None, tasks=0, deps=[],
-           threads=0, memory=8 * 1024, retries=5, requirements=[], environment={},
-           logdir=None, log=None, workdir=None, verbose=1):
-    if deps:
-        raise NotImplementedError("Cannot submit individual HTCondor jobs with dependencies yet, this requires use of DAGMan")
-    if logdir or log:
-        if not logdir:
-            log = os.path.abspath(log)
-            logdir = os.path.dirname(log)
-        elif not log:
-            logdir = os.path.abspath(logdir)
-            if tasks > 0:
-                log = os.path.join(logdir, name + "_$(Cluster).$(Process).log")
-            else:
-                log = os.path.join(logdir, name + "_$(Cluster).log")
-        makedirs(logdir)
-    jobdesc = "universe = vanilla\n"
-    if threads > 0:
-        jobdesc += "request_cpus = {0}\n".format(threads)
-    if memory > 0:
-        jobdesc += "request_memory = {0}\n".format(memory)
-    if requirements:
-        jobdesc += "requirements = " + " && ".join(requirements) + "\n"
-    if environment:
-        jobdesc += "environment = \""
-        for envname, envval in environment.items():
-            jobdesc += " {0}='{1}'".format(envname, ':'.join(envval) if isinstance(envval, (list, tuple)) else envval)
-        jobdesc += "\"\n"
-    if workdir:
-        jobdesc += "initialdir = {0}\n".format(os.path.abspath(workdir))
-    # Note: MIRTK executables return exit code 6 when memory allocation fails, other codes are kill/term signals
-    jobdesc += "on_exit_remove = (ExitBySignal == False && ExitCode != 6 && ExitCode != 247 && ExitCode != 241) || (ExitBySignal == True && ExitSignal != 9 && ExitSignal != 15)\n"
-    if retries > 0:
-        jobdesc += "max_retries = {0}\n".format(retries)
-    if script:
-        if command:
-            raise ValueError("Keyword arguments 'command' and 'script' are mutually exclusive")
-        if not log:
-            raise ValueError("Script submission of batch to HTCondor requires log path for script file!")
-        script_path = os.path.join(logdir, name + ".py")
-        with open(script_path, "wt") as f:
-            f.write(script.format(**opts))
-        jobdesc += "executable = {0}\n".format(sys.executable)
-        jobdesc += "arguments = \"'{0}'".format(script_path)
-        if tasks > 0:
-            jobdesc += " $(Process)"
-        jobdesc += "\"\n"
-        if log:
-            jobdesc += "output = {0}\n".format(log)
-            jobdesc += "error = {0}\n".format(log)
-        jobdesc += "queue"
-        if tasks > 0:
-            jobdesc += " {}".format(tasks)
-        jobdesc += "\n"
-    else:
-        jobdesc += "executable = {0}\n".format(sys.executable)
-        jobdesc += "arguments = \"-c 'import sys; import socket;"
-        jobdesc += " sys.stdout.write(\"\"Host: \"\" + socket.gethostname() + \"\"\\n\\n\"\");"
-        jobdesc += " sys.path.insert(0, \"\"{0}\"\");".format(os.path.dirname(os.path.dirname(mirtk.__file__)))
-        jobdesc += " import mirtk; mirtk.check_call([\"\"{0}\"\"] + sys.argv[1:])'".format(command if command else name)
-        for arg in args:
-            arg = str(arg)
-            if ' ' in arg:
-                arg = "'" + arg + "'"
-            jobdesc += ' ' + arg
-        for opt in opts:
-            arg = opts[opt]
-            if opt[0] != '-':
-                opt = '-' + opt
-            jobdesc += ' ' + opt
-            if arg is not None:
-                if isinstance(arg, (list, tuple)):
-                    arg = ' '.join([str(x) for x in arg])
-                else:
-                    arg = str(arg)
-                jobdesc += ' ' + arg
-        jobdesc += "\"\n"
-        if log:
-            jobdesc += "output = {0}\n".format(log)
-            jobdesc += "error = {0}\n".format(log)
-        jobdesc += "queue\n"
-    proc = subprocess.Popen(["condor_submit", "-"], stdout=subprocess.PIPE, stderr=subprocess.PIPE, stdin=subprocess.PIPE)
-    (out, err) = proc.communicate(input=jobdesc.encode('utf-8'))
-    if proc.returncode != 0:
-        raise Exception(err)
-    match = re.search('[0-9]+ job(s) submitted to cluster ([0-9]+)\.', out)
-    if not match:
-        match = re.search('\*\* Proc ([0-9]+)(\.[0-9]+)?:', out)
-        if not match:
-            raise Exception("Failed to determine job ID from condor_submit output:\n" + out)
-    jobid = int(match.group(1))
-    if verbose > 0:
-        if tasks > 0:
-            print("  Submitted batch {} (JobId={}, Tasks={})".format(name, jobid, tasks))
-        else:
-            print("  Submitted job {} (JobId={})".format(name, jobid))
-    return jobid if tasks == 0 else (jobid, tasks)
-
-
-# ----------------------------------------------------------------------------
-def parse_condor_xml(s):
-    """Parse XML ClassAds returned by condor_q or condor_history with -xml option."""
-    # Note: condor_history may return multiple (only one non-empty) <classads> tags
-    from xml.etree import ElementTree
-    end = 0
-    classads = None
-    starttag = "<classads>"
-    endtag = "</classads>"
-    while True:
-        start = s.find(starttag, end)
-        if start == -1:
-            break
-        end = s.find(endtag, start)
-        if end == -1:
-            raise ValueError("Malformed <classads> XML, could not find matching </classads>!")
-        end += len(endtag)
-        elem = ElementTree.fromstring(s[start:end])
-        if classads is None:
-            classads = elem
-        else:
-            classads.extend(elem)
-    return classads
-
-
-# ----------------------------------------------------------------------------
-def cwait(jobs, max_time=0, max_error=5, interval=60, verbose=0):
-    if not isinstance(jobs, list):
-        jobs = [jobs]
-    clusters = []
-    for job in jobs:
-        if isinstance(job, tuple):
-            cluid = job[0]
-            tasks = job[1]
-        else:
-            cluid = job
-            tasks = 1
-        if cluid > 0:
-            clusters.append((cluid, tasks))
-    num_wait = len(clusters)
-    num_error = 0
-    total_time = 0
-    iterations = 0
-    while num_wait > 0 and (max_time <= 0 or total_time < max_time or total_time == 0) and num_error < max_error:
-        time.sleep(interval)
-        total_time += interval
-        try:
-            num_pending = 0
-            num_running = 0
-            num_suspended = 0
-            num_held = 0
-            num_done = 0
-            for cluster, tasks in clusters:
-                classads = subprocess.check_output(["condor_q", "-xml", str(cluster)], stderr=subprocess.STDOUT)
-                classads = parse_condor_xml(classads)
-                for process in range(tasks):
-                    classad = classads.find(".c/a[@n='ClusterId'][i='{0}']/../a[@n='ProcId'][i='{1}']/..".format(cluster, process))
-                    if classad is None:
-                        num_done += 1
-                    else:
-                        status = int(classad.find("a[@n='JobStatus']/i").text)
-                        # 1) Idle
-                        # 2) Running
-                        # 3) Removed
-                        # 4) Completed (also when failed, check ExitCode afterwards using condor_history)
-                        # 5) Held
-                        # 6) Transferring Output
-                        # 7) Suspended
-                        if status == 1:
-                            num_pending += 1
-                        elif status == 2 or status == 6:
-                            num_running += 1
-                        elif status == 3:
-                            num_done += 1
-                        elif status == 4:
-                            num_done += 1
-                        elif status == 5:
-                            num_held += 1
-                        elif status == 7:
-                            num_suspended += 1
-                        else:
-                            raise Exception("Unknown job status: {}".format(status))
-            num_wait = num_running + num_pending + num_held + num_suspended
-            if verbose > 0 and (num_wait <= 0 or iterations % verbose == 0):
-                sys.stdout.write("{:%Y-%b-%d %H:%M:%S}".format(datetime.now()))
-                sys.stdout.write(" WAIT {p} pending, {r} running, {s} suspended, {h} held, {d} completed\n".format(
-                    p=num_pending, r=num_running, s=num_suspended, h=num_held, d=num_done
-                ))
-                sys.stdout.flush()
-            num_error = 0
-        except subprocess.CalledProcessError:
-            sys.stdout.write("{:%Y-%b-%d %H:%M:%S}".format(datetime.now()))
-            sys.stdout.write(" WAIT Failed to retrieve job status, will retry {0} more times!\n".format(max_error - num_error))
-            sys.stdout.flush()
-            num_error += 1
-        iterations += 1
-    if num_error >= max_error:
-        raise Exception("Exceeded maximum number of retries to query status of jobs!")
-    if num_wait > 0 and max_time > 0 and total_time >= max_time:
-        raise Exception("Exceeded maximum time waiting for jobs to complete!")
-    num_fail = 0
-    if total_time > 0:
-        time.sleep(10)
-        num_jobs = 0
-        unknown = {}
-        for cluster, tasks in clusters:
-            num_jobs += tasks
-            unknown[cluster] = [0] * tasks
-        num_error = 0
-        num_unknown = num_jobs
-        while num_unknown > 0 and num_error <= max_error:
-            try:
-                num_fail = 0
-                for cluster, tasks in clusters:
-                    classads = parse_condor_xml(subprocess.check_output(["condor_history", "-xml", str(cluster)]))
-                    for process in range(tasks):
-                        classad = classads.find(".c/a[@n='ClusterId'][i='{0}']/../a[@n='ProcId'][i='{1}']/..".format(cluster, process))
-                        if classad is None:
-                            unknown[cluster][process] += 1
-                        else:
-                            unknown[cluster][process] = 0
-                            status = int(classad.find("a[@n='JobStatus']/i").text)
-                            exit_code = int(classad.find("a[@n='ExitCode']/i").text)
-                            if status != 4 or exit_code != 0:
-                                num_fail += 1
-                num_unknown = 0
-                for cluster, tasks in clusters:
-                    for process in range(tasks):
-                        if unknown[cluster][process] > 0:
-                            if unknown[cluster][process] > max_error:
-                                raise Exception("Could not retrieve exit code of job {}.{} for the past {} attempts!".format(cluster, process, unknown[cluster][process]))
-                            num_unknown += 1
-                if verbose > 0:
-                    sys.stdout.write("{:%Y-%b-%d %H:%M:%S}".format(datetime.now()))
-                    sys.stdout.write(" DONE {0} succeeded, {1} failed".format(num_jobs - num_fail - num_unknown, num_fail))
-                    if num_unknown > 0:
-                        sys.stdout.write(", {} unknown, will retry".format(num_unknown))
-                    sys.stdout.write("\n")
-                    sys.stdout.flush()
-            except subprocess.CalledProcessError:
-                sys.stdout.write("{:%Y-%b-%d %H:%M:%S}".format(datetime.now()))
-                sys.stdout.write(" WAIT Failed to retrieve exit codes, will retry {0} more times!\n".format(max_error - num_error))
-                sys.stdout.flush()
-                num_error += 1
-        if num_error >= max_error:
-            raise Exception("Exceeded maximum number of retries to query exit codes of jobs!")
-    return num_fail == 0
-
-
-##############################################################################
 # Auxiliaries
 
 # ----------------------------------------------------------------------------
@@ -1938,17 +1529,6 @@ def basename_without_ext(path):
         name = os.path.splitext(name)[0]
     return name
 
-
-# ----------------------------------------------------------------------------
-def makedirs(path):
-    """Make directories, throws no error when it already exists."""
-    if not path:
-        raise ValueError("Path argument is empty or None!")
-    try:
-        os.makedirs(path)
-    except OSError as e:
-        if e.errno != errno.EEXIST:
-            raise
 
 # ----------------------------------------------------------------------------
 def remove_or_makedirs(path):
