+++ Applications/lib/python/mirtk/utils.py
@@ -1,0 +1,35 @@
+##############################################################################
+# Medical Image Registration ToolKit (MIRTK)
+#
+# Copyright 2017 Imperial College London
+# Copyright 2017 Andreas Schuh
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+##############################################################################
+
+"""Miscellaneous utility functions."""
+
+import os
+import errno
+
+
+# ----------------------------------------------------------------------------
+def makedirs(path):
+    """Make directories, throws no error when it already exists."""
+    if not path:
+        raise ValueError("Path argument is empty or None!")
+    try:
+        os.makedirs(path)
+    except OSError as e:
+        if e.errno != errno.EEXIST:
+            raise
